requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
matplotlib>=3.5.0
scipy>=1.7.0
tqdm>=4.62.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.5
lpips>=0.1.4
pyrender>=0.1.45
trimesh>=3.9.0
kornia>=0.6.0
einops>=0.6.0
wandb>=0.13.0
tensorboard>=2.10.0 